refactor: moved skills to /data
